now user can add and remove from the saved movie list

+++ BestMovies.Api/Repositories/Impl/SavedMoviesRepository.cs
@@ -69,6 +69,9 @@ public class SavedMoviesRepository : ISavedMoviesRepository
         
         if (movieToDelete is not null)
         {
+            // Only delete the movies that are not watched yet
+            if (movieToDelete.IsWatched) return;
+            
             _dbContext.SavedMovies.Remove(movieToDelete);
             await _dbContext.SaveChangesAsync();
         }
